Bug 1080491 - send "answered-elsewhere" state.
In order to implement that, each session stores if it is connected or not, and
discards the "connection" messages sent trough the pubsub in case they are.

All the other ones (the ones which aren't connected) will receive an update
with "terminated" and a reason of "answered-elsewhere", as the spec specifies.