Bug 1080491 — Fix "answered-elsewhere" impl.
Previous implementation of this was actually only sending the
"answered-elsewhere" when the call was "connected", and not when it was
transitioning from "alerting" to "accepted".

In order to make this work, I had to also to count the number of connected
devices for each session type (caller, callee), because the clients are
expecting to get their socket connection closed in case there is only one
device on the other side of the call if it closes the connection.