Bug 1099144 — Use redis hashes & sets for SP urls.
Uses redis hashes and sets rather than normal keys and doing a keys with a "*"
in the lookup part. Doing so costs a lot to redis and it uses a lot of CPU to
compare the keys before returning the values.

Using hashes and sets reduces a lot the complexity of the calls (going from
O(Number of keys in the database) to O(Size of the hash) + O(Number of sessions
per user).